feat(settings): UI fix and functionality fix

+++ apps/settings/lib/src/features/display/data/display_repository_impl.dart
@@ -13,7 +13,7 @@ class DisplayRepositoryImpl extends DisplayRepository {
   }
 
   @override
-  setBrightness(int brightness) {
+  setBrightness(double brightness) {
     return _dBusDisplayService.setBrightness(brightness);
   }
